Refactor utils and service classes, upgrade library version
The Utils class has been refactored to include the methods 'toClientDto' and 'changeFormatDate', which were previously in the ClientService class. This leads to a cleaner and more understandable separation of concerns. The pom.xml file has also been updated to use version 5.9 of the opencsv library instead of version 5.7.1. The artifactId in the pom.xml file has been changed from 'clientes' to 'clients' for better consistency with English naming conventions.